requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alpaca==1.0.0
alpaca-py==0.7.2
alpaca-trade-api==3.2.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backtrader==1.9.76.123
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colcon-core==0.12.1
colcon-pkg-config==0.1.0
comm==0.1.2
config==0.5.1
contourpy==1.0.7
coverage==7.2.1
cryptocmd==0.6.1
cryptography==39.0.2
cycler==0.11.0
dateparser==1.1.7
debugpy==1.6.6
decorator==5.1.1
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
empy==3.3.4
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.39.0
frozendict==2.3.5
frozenlist==1.3.3
greenlet==2.0.2
html5lib==1.1
idna==3.4
imageio==2.26.0
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kucoin-python==1.0.23
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.3
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.2
pytest-cov==4.0.0
pytest-repeat==0.9.1
pytest-rerunfailures==11.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-kucoin==2.1.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
schedule==1.1.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
Solution==5.4.0
soupsieve==2.4
SQLAlchemy==2.0.5.post1
sseclient-py==1.7.2
stack-data==0.6.2
tablib==3.3.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
tradingview-ta==3.3.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
vectorbt==0.24.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
yarl==1.8.2
yfinance==0.2.12
